+++ packages/experiments/src/components/Tile/Tile.tsx
@@ -1,6 +1,6 @@
 
 import * as React from 'react';
-import { ITileProps } from './Tile.Props';
+import { ITileProps, TileSize } from './Tile.Props';
 import { Check } from 'office-ui-fabric-react/lib/Check';
 import { SELECTION_CHANGE } from 'office-ui-fabric-react/lib/Selection';
 import { css, BaseComponent, autobind, getId } from 'office-ui-fabric-react/lib/Utilities';
@@ -15,14 +15,42 @@ const SignalStyles: any = SignalStylesModule;
 
 const enum TileLayoutValues {
   nameplatePadding = 12,
-  nameplateNameHeight = 20,
-  nameplateActivityHeight = 20,
+  largeNameplateNameHeight = 17,
+  smallNameplateNameHeight = 14,
+  nameplateMargin = 4,
+  largeNameplateActivityHeight = 14,
+  smallNameplateActivityHeight = 14,
   foregroundMargin = 16
 }
 
 export interface ITileState {
   isSelected?: boolean;
 }
+
+const SIZES: {
+  [P in TileSize]: {
+    nameplatePadding: number;
+    nameplateNameHeight: number;
+    nameplateMargin: number;
+    nameplateActivityHeight: number;
+    foregroundMargin: number;
+  };
+} = {
+    small: {
+      nameplatePadding: TileLayoutValues.nameplatePadding,
+      nameplateNameHeight: TileLayoutValues.smallNameplateNameHeight,
+      nameplateMargin: TileLayoutValues.nameplateMargin,
+      nameplateActivityHeight: TileLayoutValues.smallNameplateActivityHeight,
+      foregroundMargin: TileLayoutValues.foregroundMargin
+    },
+    large: {
+      nameplatePadding: TileLayoutValues.nameplatePadding,
+      nameplateNameHeight: TileLayoutValues.largeNameplateNameHeight,
+      nameplateMargin: TileLayoutValues.nameplateMargin,
+      nameplateActivityHeight: TileLayoutValues.largeNameplateActivityHeight,
+      foregroundMargin: TileLayoutValues.foregroundMargin
+    }
+  };
 
 /**
  * A tile provides a frame for a potentially-selectable item which displays its contents prominently.
@@ -113,10 +141,14 @@ export class Tile extends BaseComponent<ITileProps, ITileState> {
       foreground,
       showBackgroundFrame = false,
       showForegroundFrame = false,
+      hideBackground = false,
+      hideForeground = false,
       itemName,
       itemActivity,
       componentRef,
       className,
+      tileSize = 'large',
+      contentSize,
       ...aProps
     } = this.props;
 
@@ -133,10 +165,14 @@ export class Tile extends BaseComponent<ITileProps, ITileState> {
         aria-labelledby={ this._nameId }
         aria-describedby={ this._activityId }
         className={ css('ms-Tile', className, TileStyles.tile, {
+          [`ms-Tile--isSmall ${TileStyles.isSmall}`]: tileSize === 'small',
+          [`ms-Tile--isLarge ${TileStyles.isLarge}`]: tileSize === 'large',
           [`ms-Tile--hasBackgroundFrame ${TileStyles.hasBackgroundFrame}`]: showBackgroundFrame,
           [`ms-Tile--isSelected ${TileStyles.selected} ${SignalStyles.selected}`]: isSelected,
           [`ms-Tile--isSelectable ${TileStyles.selectable}`]: isSelectable,
-          [`ms-Tile--hasBackground ${TileStyles.hasBackground} ${SignalStyles.dark}`]: !!background
+          [`ms-Tile--hasBackground ${TileStyles.hasBackground}`]: !!background,
+          [SignalStyles.dark]: !!background && !hideBackground,
+          [`ms-Tile--showBackground ${TileStyles.showBackground}`]: !hideBackground
         }) }
         data-is-focusable={ true }
         data-is-sub-focuszone={ true }
@@ -150,13 +186,15 @@ export class Tile extends BaseComponent<ITileProps, ITileState> {
         >
           {
             background ? this._onRenderBackground({
-              background: background
+              background: background,
+              hideBackground
             }) : null
           }
           {
             foreground ? this._onRenderForeground({
               foreground: foreground,
-              showForegroundFrame: showForegroundFrame
+              showForegroundFrame: showForegroundFrame,
+              hideForeground
             }) : null
           }
           {
@@ -176,12 +214,18 @@ export class Tile extends BaseComponent<ITileProps, ITileState> {
   }
 
   private _onRenderBackground({
-    background
+    background,
+    hideBackground
   }: {
-      background: React.ReactNode | React.ReactNode[]
+      background: React.ReactNode | React.ReactNode[];
+      hideBackground: boolean;
     }): JSX.Element {
     return (
-      <span className={ css('ms-Tile-background', TileStyles.background) }>
+      <span
+        className={ css('ms-Tile-background', TileStyles.background, {
+          [`ms-Tile-background--hide ${TileStyles.backgroundHide}`]: hideBackground
+        }) }
+      >
         { background }
       </span>
     );
@@ -189,24 +233,27 @@ export class Tile extends BaseComponent<ITileProps, ITileState> {
 
   private _onRenderForeground({
     foreground,
-    showForegroundFrame
+    showForegroundFrame,
+    hideForeground
   }: {
       foreground: React.ReactNode | React.ReactNode[];
       showForegroundFrame: boolean;
+      hideForeground: boolean;
     }): JSX.Element {
     return (
       <span
         role='presentation'
-        className={ css('ms-Tile-content', TileStyles.content) }
+        className={ css('ms-Tile-aboveNameplate', TileStyles.aboveNameplate) }
       >
         <span
           role='presentation'
-          className={ css('ms-Tile-foreground', TileStyles.foreground) }
+          className={ css('ms-Tile-content', TileStyles.content) }
         >
           <span
             role='presentation'
-            className={ css('ms-Tile-frame', TileStyles.frame, {
-              [`ms-Tile-frame--hasForegroundFrame ${TileStyles.hasForegroundFrame}`]: showForegroundFrame
+            className={ css('ms-Tile-foreground', TileStyles.foreground, {
+              [`ms-Tile-foreground--hasFrame ${TileStyles.hasForegroundFrame}`]: showForegroundFrame,
+              [`ms-Tile-foreground--hide ${TileStyles.foregroundHide}`]: hideForeground
             }) }
           >
             { foreground }
@@ -292,7 +339,8 @@ export function getTileLayout(tileElement: JSX.Element): ITileLayout {
   const tileProps: ITileProps = tileElement.props;
 
   const {
-    contentSize
+    contentSize,
+    tileSize = 'large'
   } = tileProps;
 
   if (!contentSize) {
@@ -301,22 +349,30 @@ export function getTileLayout(tileElement: JSX.Element): ITileLayout {
 
   const width = contentSize.width;
 
+  const {
+    nameplatePadding,
+    nameplateMargin,
+    nameplateActivityHeight,
+    nameplateNameHeight,
+    foregroundMargin
+  } = SIZES[tileSize];
+
   let nameplateHeight = 0;
 
   if (tileProps.itemName || tileProps.itemActivity) {
-    nameplateHeight += TileLayoutValues.nameplatePadding * 2; // 12px top/bottom padding.
+    nameplateHeight += nameplatePadding * 2; // 12px top/bottom padding.
     if (tileProps.itemName) {
-      nameplateHeight += TileLayoutValues.nameplateNameHeight;
+      nameplateHeight += nameplateNameHeight;
     }
     if (tileProps.itemActivity) {
-      nameplateHeight += TileLayoutValues.nameplateActivityHeight;
+      nameplateHeight += nameplateActivityHeight + nameplateMargin;
     }
   }
 
   return {
     foregroundSize: {
-      width: width - TileLayoutValues.foregroundMargin * 2,
-      height: contentSize.height - TileLayoutValues.foregroundMargin - nameplateHeight
+      width: width - foregroundMargin * 2,
+      height: contentSize.height - foregroundMargin - nameplateHeight
     },
     backgroundSize: contentSize
   };