+++ packages/office-ui-fabric-react/src/components/GroupedList/examples/GroupedList.Custom.Example.tsx
@@ -60,7 +60,7 @@ export class GroupedListCustomExample extends React.Component<any, any> {
 
   private _onRenderFooter(props: IGroupDividerProps): JSX.Element {
     return (
-      <div className={ css('ms-GroupedListExample-footer', FontClassNames.xLarge) }>
+      <div className={ css('ms-GroupedListExample-footer', FontClassNames.large) }>
         This is a custom footer for { props.group!.name }
       </div>
     );