+++ packages/office-ui-fabric-react/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - office-ui-fabric-react
 
-This log was last generated on Tue, 05 Sep 2017 20:36:48 GMT and should not be manually modified.
+This log was last generated on Thu, 07 Sep 2017 10:09:51 GMT and should not be manually modified.
+
+## 4.42.0
+Thu, 07 Sep 2017 10:09:51 GMT
+
+### Minor changes
+
+- Added rootExpandedHovered and iconExpandedHovered to IButtonStyles interface.  These entries are for specifying on hover style for expanded state.
+- ContextualMenu: `onRenderSubMenu` prop added to allow the overriding of submenu rendering.
+
+## 4.41.0
+Wed, 06 Sep 2017 10:16:32 GMT
+
+### Minor changes
+
+- Allow rendering empty GroupedList headers
 
 ## 4.40.1
 Tue, 05 Sep 2017 20:36:48 GMT